Refactor module and homework management; update validation logic, error handling, and improve response messages

+++ 06 Node Diplomate Final/src/controllers/year.controller.js
@@ -1,12 +1,22 @@
 import ResponseBuilder from "../utils/builders/responseBuilder.builder.js";
 import * as db from "../dataBase/models.js";
 import YearRepository from "../repositories/year.repository.js";
+import modulesValidations from "../utils/modulesValidation.util.js";
 
 // ~ ------------------------------------> Create year <------------------------------------ ~
 const createYear = async (req, res) => {
   const { year, careerId } = req.body;
 
   try {
+     // ^ --------------> Enviar los datos a la función de validación
+     const Validations = modulesValidations({year});
+
+     // ^ --------------> Validar si hay errores
+     if(Validations.getOk() === false){
+       console.error('[Modules.Controller.Create] - Validation error');
+       return res.status(400).json(Validations.response);
+     }
+
     //! ---> Si la carrera no existe, lanzar un error
     if (!careerId) {
       const response = new ResponseBuilder().setOk(false).setStatus(400).setMessage("Career is required").build();
@@ -15,13 +25,6 @@ const createYear = async (req, res) => {
       return res.status(400).json(response);
     }
 
-    //! ---> Si no se envía el año o no es un numero, lanzar un error
-    if (!year || isNaN(year)) {
-      const response = new ResponseBuilder().setOk(false).setStatus(400).setMessage("Year is required and must be a number").build();
-      // Enviar respuesta
-      console.error("[Year.Controller.Create] - Year is required and must be a number");
-      return res.status(400).json(response);
-    }
 
     //^ -----> Creamos el año y lo guardamos
     const savedYear = await YearRepository.createYear(year, careerId);
@@ -61,6 +64,23 @@ const deleteYear = async (req, res) => {
   const { year, careerId } = req.body;
 
   try {
+    //! ---> Si no se envia la carrera, lanzar un error
+    if (!careerId) {
+      const response = new ResponseBuilder().setOk(false).setStatus(400).setMessage("Career is required").build();
+      // Enviar respuesta
+      console.error("[Year.Controller.Delete] - Career is required");
+      return res.status(400).json(response);
+    }
+
+    // ^ --------------> Enviar los datos a la función de validación
+    const Validations = modulesValidations({year});
+
+    // ^ --------------> Validar si hay errores
+    if(Validations.getOk() === false){
+      console.error('[Modules.Controller.Create] - Validation error');
+      return res.status(400).json(Validations.response);
+    }    
+
     // Eliminar el año
     const deletedYear = await YearRepository.removeYear(year, careerId);
 
@@ -106,6 +126,14 @@ const getAllYears = async (req, res) => {
   const { careerId } = req.body;
 
   try {
+    //! ---> Si no se envía la carrera, lanzar un error
+    if (!careerId) {
+      const response = new ResponseBuilder().setOk(false).setStatus(400).setMessage("Career is required").build();
+      // Enviar respuesta
+      console.error("[Year.Controller.GetAll] - Career is required");
+      return res.status(400).json(response);
+    }
+
     // Obtener todos los años de una carrera
     const years = await YearRepository.getAllYears(careerId);
